add useless comment

+++ math_api/math_api/math/operations.py
@@ -8,7 +8,7 @@ def sub(t1: int, t2: int) -> int:
     return t1 - t2
 
 
-def div(t1: int, t2: int) -> int:
+def div(t1: int, t2: int) -> float:
     """For two integers t1 and t2, returns t1/t2."""
     if t2 == 0:
         raise ValueError("Division by 0")
